+++ doc/release_notes.rst
@@ -28,6 +28,8 @@ E.g. if a new rule becomes available describe how to use it `make test` and in o
 
 * Fix the mismatch between buses and x, y locations while creating H2 Stores `PR #1134 <https://github.com/pypsa-meets-earth/pypsa-earth/pull/1134>`_
 
+* Enable configfile specification for mock_snakemake `PR #1135 <https://github.com/pypsa-meets-earth/pypsa-earth/pull/1135>`_
+
 PyPSA-Earth 0.4.1
 =================